# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Page": {
        "description": "A page of a listing: the items and the cursor to ask for the next one.",
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "cursor of the next page, null at the end",
            "title": "Next"
          }
        },
        "title": "Page",
        "type": "object"
      },
      "Problem": {
        "description": "An error as RFC 9457 describes it: the same shape whatever went wrong.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "what happened in this occurrence",
            "title": "Detail"
          },
          "instance": {
            "description": "path where it happened",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "short human summary, the same for every occurrence",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "identifier of the kind of problem",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "instance"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "ReviewDecision": {
        "description": "What a DPO reviewer decides about a column the classifier could not settle.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "only when correcting",
            "title": "Category"
          },
          "decision": {
            "description": "accept confirms the proposed category; reject leaves the column unclassified; correct rejects the proposal and classifies the column as `category`",
            "enum": [
              "accept",
              "reject",
              "correct"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "access token issued by the appliance Keycloak",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Campaigns, inventory, findings and evidence of the ARGOS appliance. Listings paginate by opaque cursor, errors follow RFC 9457 and every POST that creates accepts an Idempotency-Key.",
    "title": "ARGOS API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/inventory/coverage": {
      "get": {
        "operationId": "coverage_api_v1_inventory_coverage_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Coverage Api V1 Inventory Coverage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Coverage, freshness and pending review of the inventory",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/nodes/{node_key}": {
      "get": {
        "operationId": "node_api_v1_inventory_nodes__node_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "node_key",
            "required": true,
            "schema": {
              "title": "Node Key",
              "type": "string"
            }
          },
          {
            "description": "neighbours to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "neighbours to return",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Node Api V1 Inventory Nodes  Node Key  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "A node of the inventory graph",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/review-queue": {
      "get": {
        "operationId": "review_queue_api_v1_inventory_review_queue_get",
        "parameters": [
          {
            "description": "opaque cursor returned by the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "opaque cursor returned by the previous page",
              "title": "Cursor"
            }
          },
          {
            "description": "items per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "items per page",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Columns awaiting human review",
        "tags": [
          "inventory"
        ]
      }
    },
    "/api/v1/inventory/review-queue/{node_key}": {
      "post": {
        "operationId": "review_api_v1_inventory_review_queue__node_key__post",
        "parameters": [
          {
            "in": "path",
            "name": "node_key",
            "required": true,
            "schema": {
              "title": "Node Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Review Api V1 Inventory Review Queue  Node Key  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "no valid token"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the role does not allow it"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "it does not exist"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "the state does not allow it"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "invalid body"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "not configured yet"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide about a column under review",
        "tags": [
          "inventory"
        ]
      }
    },
...
  }
}